backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
sentence-transformers
transformers
torch
pandas
scikit-learn
tqdm
requests
PyPDF2
openai
ollama
psycopg2-binary
hf_xet
supabase